requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0
pydantic>=2.6,<3.0
numpy>=1.24
scikit-learn>=1.3
sentence-transformers>=3.0
openai>=1.30
python-dotenv>=1.0
pytest>=8.0
